# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Battle Rap API
  version: 1.0.0
  description: >
    REST API for managing online rap battles, tournaments, submissions, judging,
    and administration workflows.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/battles/{id}:
    get:
      tags: [Public]
      summary: Get battle summary
      operationId: getBattle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Battle details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Battle'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/media/presign:
    post:
      tags: [Media]
      summary: Request a presigned upload URL
      operationId: createPresignedUpload
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignRequest'
      responses:
        '200':
          description: Presigned URL details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/media/complete:
    post:
      tags: [Media]
      summary: Mark upload as completed
      operationId: completeUpload
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteUploadRequest'
      responses:
        '200':
          description: Media asset recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAssetStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/media/{assetId}/status:
    get:
      tags: [Media]
      summary: Get media asset status
      operationId: getMediaAssetStatus
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: assetId
          required: true
          schema:
            type: string
            format: uuid
          description: Media asset identifier.
      responses:
        '200':
          description: Media asset status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAssetStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/profile/me:
    get:
      tags: [Profile]
      summary: Get current user profile
      operationId: getOwnProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Authenticated user profile view.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/profile/{id}:
    get:
      tags: [Profile]
      summary: Get public profile by user ID
      operationId: getProfileById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User identifier.
      responses:
        '200':
          description: Profile view adapted to caller visibility.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileView'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/profile/{id}/highlights:
    get:
      tags: [Profile]
      summary: Get battle statistics and history for a profile
      operationId: getProfileHighlights
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User identifier.
      responses:
        '200':
          description: Aggregated statistics and battle lists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileHighlights'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/judge/assignments:
    get:
      tags: [Judge]
      summary: List judge assignments
      operationId: listJudgeAssignments
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Assignments for the judge.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JudgeAssignment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/judge/assignments/random:
    post:
      tags: [Judge]
      summary: Request next battle assignment
      operationId: requestJudgeAssignment
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Newly assigned battle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgeAssignment'
        '204':
          description: No available battles to assign.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/judge/assignments/{id}/status:
    post:
      tags: [Judge]
      summary: Update judge assignment status
      operationId: updateJudgeAssignmentStatus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [completed, skipped]
      responses:
        '200':
          description: Updated assignment state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgeAssignment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/judge/battles/{id}:
    get:
      tags: [Judge]
      summary: Get battle details for judging
      operationId: getJudgeBattleDetails
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Battle information for judge.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgeBattleDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/judge/battles/{id}/scores:
    post:
      tags: [Judge]
      summary: Submit or update evaluation
      operationId: upsertEvaluation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationRequest'
      responses:
        '201':
          description: Evaluation saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token issued by the API. Required for authenticated endpoints.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        Refresh token cookie issued during login/registration.
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication required or token invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        status:
          type: integer
        field_errors:
          type: object
          additionalProperties:
            type: string
      example:
        type: https://httpstatuses.com/422
        title: Payload validation failed.
        detail: Payload validation failed.
        status: 422
        field_errors:
          email: Invalid email.
...
    Battle:
      type: object
      required:
        - id
        - round_id
        - status
      properties:
        id:
          type: string
          format: uuid
        round_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        ends_at:
          type: string
          format: date-time
          nullable: true
        winner_match_track_id:
          type: string
          format: uuid
          nullable: true
        engagement:
          $ref: '#/components/schemas/BattleEngagementSummary'
...
    BattleEngagementSummary:
      type: object
      required:
        - comments
      properties:
        comments:
          type: integer
          minimum: 0
...
    ProfileAvatar:
      type: object
      required:
        - key
        - url
      properties:
        key:
          type: string
          description: Internal storage key for the avatar asset.
        url:
          type: string
          format: uri
          description: CDN URL for the avatar asset.
    ProfileViewerContext:
      type: object
      required:
        - is_self
        - can_edit
        - can_moderate
        - can_view_private
      properties:
        is_self:
          type: boolean
          description: Indicates the profile belongs to the caller.
        can_edit:
          type: boolean
          description: Caller may submit profile change requests.
        can_moderate:
          type: boolean
          description: Caller has moderator or admin privileges over this profile.
        can_view_private:
          type: boolean
          description: Caller may see private fields such as email or full name.
    ProfileView:
      type: object
      required:
        - id
        - display_name
        - roles
        - created_at
        - updated_at
        - viewer_context
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        roles:
          type: array
          items:
            type: string
        avatar:
          $ref: '#/components/schemas/ProfileAvatar'
          nullable: true
        bio:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        viewer_context:
          $ref: '#/components/schemas/ProfileViewerContext'
        email:
          type: string
          format: email
          description: Present when the caller can view private information.
        age:
          type: integer
          minimum: 12
          maximum: 120
          nullable: true
        vk_id:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        socials:
          type: object
          additionalProperties: true
          description: Social handles map, available to privileged viewers.
    ProfileBattleParticipant:
      type: object
      required:
        - user_id
        - display_name
        - is_self
      properties:
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
          nullable: true
        is_self:
          type: boolean
        seed:
          type: integer
          nullable: true
    ProfileBattleRoundInfo:
      type: object
      required:
        - id
        - number
        - kind
        - status
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
        kind:
          type: string
        status:
          type: string
    ProfileBattleTournamentInfo:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
    ProfileBattle:
      type: object
      required:
        - id
        - status
        - round
        - tournament
        - participants
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        starts_at:
          type: string
          format: date-time
          nullable: true
        ends_at:
          type: string
          format: date-time
          nullable: true
        round:
          $ref: '#/components/schemas/ProfileBattleRoundInfo'
        tournament:
          $ref: '#/components/schemas/ProfileBattleTournamentInfo'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ProfileBattleParticipant'
    ProfileBattleResult:
      type: string
      enum: [win, loss, draw, pending]
    ProfileParticipantBattle:
      allOf:
        - $ref: '#/components/schemas/ProfileBattle'
        - type: object
          required:
            - result
          properties:
            result:
              $ref: '#/components/schemas/ProfileBattleResult'
    ProfileJudgedBattle:
      allOf:
        - $ref: '#/components/schemas/ProfileBattle'
        - type: object
          required:
            - evaluated_at
          properties:
            evaluated_at:
              type: string
              format: date-time
    ProfileHighlights:
      type: object
      required:
        - stats
        - participated_battles
        - judged_battles
      properties:
        stats:
          type: object
          required:
            - wins
            - losses
            - slivs
          properties:
            wins:
              type: integer
              minimum: 0
            losses:
              type: integer
              minimum: 0
            slivs:
              type: integer
              minimum: 0
        participated_battles:
          type: array
          items:
            $ref: '#/components/schemas/ProfileParticipantBattle'
        judged_battles:
          type: array
          items:
            $ref: '#/components/schemas/ProfileJudgedBattle'
...
    PresignRequest:
      type: object
      required:
        - filename
        - mime
        - size_bytes
        - type
      properties:
        filename:
          type: string
        mime:
          type: string
        size_bytes:
          type: integer
          minimum: 1
        type:
          type: string
          enum: [audio, image]
    PresignResponse:
      type: object
      required:
        - assetId
        - storageKey
        - uploadUrl
        - headers
      properties:
        assetId:
          type: string
          format: uuid
        storageKey:
          type: string
        uploadUrl:
          type: string
          format: uri
        headers:
          type: object
          additionalProperties:
            type: string
    CompleteUploadRequest:
      type: object
      required:
        - asset_id
        - storage_key
        - mime
        - size_bytes
        - kind
      properties:
        asset_id:
          type: string
          format: uuid
        storage_key:
          type: string
        mime:
          type: string
        size_bytes:
          type: integer
          minimum: 1
        kind:
          type: string
          enum: [audio, image]
    MediaAssetStatus:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          description: Current media asset status.
...
    JudgeAssignment:
      type: object
      required:
        - id
        - match_id
        - status
        - round_id
      properties:
        id:
          type: string
          format: uuid
        match_id:
          type: string
          format: uuid
        status:
          type: string
        assigned_at:
          type: string
          format: date-time
          nullable: true
        round_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
          nullable: true
        match_status:
          type: string
        round_kind:
          type: string
        round_number:
          type: integer
        round_scoring:
          type: string
        round_status:
          type: string
        round_strategy:
          type: string
          nullable: true
        judging_deadline_at:
          type: string
          format: date-time
          nullable: true
    JudgingRubricCriterion:
      type: object
      required:
        - key
        - name
      properties:
        key:
          type: string
        name:
          type: string
        weight:
          type: number
          nullable: true
        min_value:
          type: number
          nullable: true
        max_value:
          type: number
          nullable: true
        position:
          type: integer
          nullable: true
    JudgeBattleDetails:
      type: object
      required:
        - match
        - participants
        - rubric
      properties:
        match:
          type: object
          required:
            - id
            - round_id
            - status
            - round
          properties:
            id:
              type: string
              format: uuid
            round_id:
              type: string
              format: uuid
            status:
              type: string
            starts_at:
              type: string
              format: date-time
              nullable: true
            ends_at:
              type: string
              format: date-time
              nullable: true
            winner_match_track_id:
              type: string
              format: uuid
              nullable: true
            round:
              type: object
              required:
                - id
                - kind
                - number
                - scoring
                - status
                - strategy
              properties:
                id:
                  type: string
                  format: uuid
                kind:
                  type: string
                number:
                  type: integer
                scoring:
                  type: string
                  enum: [pass_fail, points, rubric]
                status:
                  type: string
                strategy:
                  type: string
                judging_deadline_at:
                  type: string
                  format: date-time
                  nullable: true
        participants:
          type: array
          items:
            type: object
            required:
              - participant_id
              - user_id
              - display_name
            properties:
              participant_id:
                type: string
                format: uuid
              user_id:
                type: string
                format: uuid
              display_name:
                type: string
              seed:
                type: integer
                nullable: true
              avg_total_score:
                type: number
                nullable: true
              track:
                type: object
                nullable: true
                properties:
                  id:
                    type: string
                    format: uuid
                  audio_key:
                    type: string
                    nullable: true
                  audio_url:
                    type: string
                    format: uri
                    nullable: true
                  mime:
                    type: string
                    nullable: true
                  duration_sec:
                    type: number
                    nullable: true
                  submitted_at:
                    type: string
                    format: date-time
                    nullable: true
                  lyrics:
                    type: string
                    nullable: true
        rubric:
          type: array
          items:
            $ref: '#/components/schemas/JudgingRubricCriterion'
        evaluation:
          type: object
          nullable: true
          properties:
            pass:
              type: boolean
              nullable: true
            score:
              type: number
              nullable: true
            rubric:
              type: object
              nullable: true
              additionalProperties:
                type: number
            total_score:
              type: number
              nullable: true
            comment:
              type: string
              nullable: true
...
    EvaluationRequest:
      type: object
      properties:
        rubric:
          type: object
          additionalProperties:
            type: number
        score:
          type: number
          minimum: 0
          maximum: 100
        pass:
          type: boolean
        comment:
          type: string
          maxLength: 2000
      description: At least one of `rubric`, `score`, or `pass` must be provided.
    EvaluationResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
          example: true